friendify/backend/requirements.txt
asgiref==3.7.2
autopep8==2.0.2
bcrypt==3.2.0
blinker==1.6.3
certifi==2022.12.7
cffi==1.14.5
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.6
django-allauth==0.54.0
django-birthday==0.1.3
django-cors-headers==4.3.0
django-rest-auth==0.9.5
django-rest-auth-forked==0.9.5
djangorestframework==3.14.0
filelock==3.12.0
Flask==3.0.0
Flask-Cors==4.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql==0.0.3
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
pipenv==2023.4.20
platformdirs==3.4.0
pycodestyle==2.10.0
pycparser==2.20
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2020.5
requests==2.31.0
requests-oauthlib==1.3.1
six==1.15.0
sqlparse==0.4.1
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.22.0
virtualenv-clone==0.5.7
Werkzeug==3.0.1
zipp==3.17.0
